requirements.txt
aiohttp==3.7.4.post0
alembic==1.7.4
amqp==2.6.1
apache-superset==1.3.2
apispec==3.3.2
async-generator==1.10
async-timeout==3.0.1
attrs==21.2.0
Authlib==0.15.5
Babel==2.9.1
backoff==1.11.1
billiard==3.6.4.0
black==21.9b0
bleach==3.3.1
Brotli==1.0.9
cachelib==0.1.1
celery==4.4.7
certifi==2021.10.8
cffi==1.15.0
chardet==4.0.0
click==7.1.2
colorama==0.4.4
convertdate==2.3.2
cron-descriptor==1.2.24
croniter==1.0.15
cryptography==35.0.0
defusedxml==0.7.1
deprecation==2.1.0
dnspython==2.1.0
email-validator==1.1.3
Flask==1.1.4
Flask-AppBuilder==3.3.4
Flask-Babel==2.0.0
Flask-Caching==1.10.1
Flask-Compress==1.10.1
Flask-JWT-Extended==3.25.1
Flask-Login==0.4.1
Flask-Migrate==3.1.0
Flask-OpenID==1.3.0
Flask-SQLAlchemy==2.5.1
flask-talisman==0.8.1
Flask-WTF==0.14.3
geographiclib==1.52
geopy==2.2.0
graphlib-backport==1.0.3
gunicorn==20.0.4
h11==0.12.0
holidays==0.10.3
humanize==3.12.0
idna==3.3
isodate==0.6.0
itsdangerous==1.1.0
Jinja2==2.11.3
jsonschema==3.2.0
kombu==4.6.11
korean-lunar-calendar==0.2.1
Mako==1.1.5
Markdown==3.3.4
MarkupSafe==2.0.1
marshmallow==3.14.0
marshmallow-enum==1.5.1
marshmallow-sqlalchemy==0.26.1
msgpack==1.0.2
multidict==5.2.0
mypy-extensions==0.4.3
numpy==1.21.3
outcome==1.1.0
packaging==21.0
pandas==1.2.5
parsedatetime==2.6
pathspec==0.9.0
pgsanity==0.2.9
Pillow==8.4.0
platformdirs==2.4.0
polyline==1.4.0
prison==0.2.1
pyarrow==4.0.1
pycparser==2.20
PyJWT==1.7.1
PyMeeus==0.5.11
pyOpenSSL==21.0.0
pyparsing==2.4.7
pyrsistent==0.18.0
python-dateutil==2.8.2
python-dotenv==0.19.1
python-geohash==0.8.5
python3-openid==3.2.0
pytz==2021.3
PyYAML==6.0
redis==3.5.3
regex==2021.10.23
selenium==4.0.0
simplejson==3.17.5
six==1.16.0
slackclient==2.5.0
sniffio==1.2.0
sortedcontainers==2.4.0
SQLAlchemy==1.3.24
SQLAlchemy-Utils==0.36.8
sqlparse==0.3.0
tabulate==0.8.9
tomli==1.2.2
trio==0.19.0
trio-websocket==0.9.2
typing-extensions==3.10.0.2
urllib3==1.26.7
vine==1.3.0
webencodings==0.5.1
Werkzeug==1.0.1
wsproto==1.0.0
WTForms==2.3.3
WTForms-JSON==0.3.3
yarl==1.7.0
